refactor(admin): extract shared gateway contracts
