+++ src/services/adminGatewayApi.ts
@@ -6,93 +6,14 @@
 import axios from '@nextcloud/axios'
 import { generateOcsUrl } from '@nextcloud/router'
 
-export interface FieldDefinition {
-	field: string
-	prompt: string
-	helper?: string
-	default: string
-	optional: boolean
-	type?: string
-	hidden?: boolean
-	min?: number
-	max?: number
-}
-
-export interface GatewayInstance {
-	id: string
-	providerId: string
-	label: string
-	default: boolean
-	createdAt: string
-	config: Record<string, string>
-	isComplete: boolean
-	groupIds: string[]
-	priority: number
-}
-
-export interface GatewayGroup {
-	id: string
-	displayName: string
-}
-
-export interface GatewayProviderDefinition {
-	id: string
-	name: string
-	fields: FieldDefinition[]
-}
-
-export interface GatewayInfo {
-	id: string
-	name: string
-	instructions: string
-	allowMarkdown: boolean
-	fields: FieldDefinition[]
-	providerSelector?: FieldDefinition
-	providerCatalog?: GatewayProviderDefinition[]
-	instances: GatewayInstance[]
-}
-
-export interface TestResult {
-	success: boolean
-	message: string
-	accountInfo?: {
-		account_name?: string
-		account_avatar_url?: string
-	}
-}
-
-export interface InteractiveSetupResponse {
-	status: 'needs_input' | 'pending' | 'done' | 'error' | 'cancelled'
-	message?: string
-	messageType?: 'info' | 'success' | 'warning' | 'error'
-	sessionId?: string
-	step?: string
-	data?: Record<string, unknown>
-	config?: Record<string, string>
-}
-
-type GatewayInstancePayload = Partial<GatewayInstance> & {
-	id: string
-	label: string
-	default: boolean
-	createdAt: string
-	config: Record<string, string>
-	isComplete: boolean
-}
-
-function normalizeInstance(instance: GatewayInstancePayload, fallbackProviderId: string): GatewayInstance {
-	return {
-		id: instance.id,
-		providerId: instance.providerId ?? fallbackProviderId,
-		label: instance.label,
-		default: instance.default,
-		createdAt: instance.createdAt,
-		config: instance.config,
-		isComplete: instance.isComplete,
-		groupIds: Array.isArray(instance.groupIds) ? instance.groupIds : [],
-		priority: typeof instance.priority === 'number' ? instance.priority : 0,
-	}
-}
+import {
+	normalizeGatewayInstance,
+	type GatewayInfo,
+	type GatewayInstancePayload,
+	type GatewayGroup,
+	type InteractiveSetupResponse,
+	type TestResult,
+} from './adminGatewayTypes.ts'
 
 function ocsData<T>(response: { data: unknown }): T {
 	const d = response.data as Record<string, Record<string, T>>
@@ -114,7 +35,7 @@ export async function listGateways(): Promise<GatewayInfo[]> {
 	const gateways = ocsData<GatewayInfo[]>(response)
 	return gateways.map((gateway) => ({
 		...gateway,
-		instances: gateway.instances.map((instance) => normalizeInstance(instance, gateway.id)),
+		instances: gateway.instances.map((instance) => normalizeGatewayInstance(instance, gateway.id)),
 	}))
 }
 
@@ -143,7 +64,7 @@ export async function createInstance(
 		{ label, config, groupIds, priority },
 	)
 	const instance = ocsData<GatewayInstancePayload>(response)
-	return normalizeInstance(instance, gatewayId)
+	return normalizeGatewayInstance(instance, gatewayId)
 }
 
 /**
@@ -157,7 +78,7 @@ export async function getInstance(gatewayId: string, instanceId: string): Promis
 		}),
 	)
 	const instance = ocsData<GatewayInstancePayload>(response)
-	return normalizeInstance(instance, gatewayId)
+	return normalizeGatewayInstance(instance, gatewayId)
 }
 
 /**
